refactor(proxy): move key resolved models helpers to model_checks
- Relocate GET /key/{id}/models payload logic from management_endpoints to litellm.proxy.auth.model_checks.
- Remove key_resolved_models_helpers.py; key_management_endpoints imports from model_checks.
- Update unit tests to import from model_checks and use _filter_key_models_by_search.

Note: pytest for this module was not run locally (no .venv); Vitest dashboard tests passed.
Made-with: Cursor
